requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
anyascii==0.3.3
    # via wagtail
asgiref==3.9.1
    # via django
beautifulsoup4==4.13.4
    # via wagtail
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
defusedxml==0.7.1
    # via willow
django==5.2.4
    # via
    #   -r requirements.in
    #   django-filter
    #   django-modelcluster
    #   django-permissionedforms
    #   django-stubs-ext
    #   django-taggit
    #   django-tasks
    #   django-treebeard
    #   djangorestframework
    #   laces
    #   wagtail
django-environ==0.12.0
    # via -r requirements.in
django-filter==25.1
    # via wagtail
django-modelcluster==6.4
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-stubs-ext==5.2.2
    # via django-tasks
django-taggit==6.1.0
    # via wagtail
django-tasks==0.7.0
    # via wagtail
django-treebeard==4.7.1
    # via wagtail
djangorestframework==3.16.0
    # via wagtail
draftjs-exporter==5.1.0
    # via wagtail
et-xmlfile==2.0.0
    # via openpyxl
filetype==1.2.0
    # via willow
idna==3.10
    # via requests
laces==0.1.2
    # via wagtail
openpyxl==3.1.5
    # via wagtail
pillow==11.3.0
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==1.0.0
    # via willow
psycopg2-binary==2.9.10
    # via -r requirements.in
requests==2.32.4
    # via wagtail
soupsieve==2.7
    # via beautifulsoup4
sqlparse==0.5.3
    # via django
telepath==0.3.1
    # via wagtail
typing-extensions==4.14.1
    # via
    #   beautifulsoup4
    #   django-stubs-ext
    #   django-tasks
urllib3==2.5.0
    # via requests
wagtail==7.0.2
    # via -r requirements.in
wheel==0.45.1
    # via -r requirements.in
willow[heif]==1.11.0
    # via
    #   wagtail
    #   willow
